Move buildScheduleDays to ProvidedService to break circular dependency.
